+++ test/src/perf.c
@@ -247,7 +247,7 @@ void test_perf_viterbi(void)
     imm_results_destroy(results);
     imm_seq_destroy(seq);
 
-    for (uint32_t i = 0; i < nmm_model_nstates(model); ++i)
+    for (uint16_t i = 0; i < nmm_model_nstates(model); ++i)
         imm_state_destroy(nmm_model_state(model, i));
 
     for (uint32_t i = 0; i < nmm_model_nbase_tables(model); ++i)
